+++ TDTChocolate/FoundationAdditions/NSFileManager+TDTAdditions.h
@@ -19,4 +19,12 @@
 - (NSURL *)tdt_fileURLToTemporaryFileWithNamePrefix:(NSString *)prefix
                                              suffix:(NSString *)suffix;
 
+/**
+ Extend the @p moveItemAtPath: method to remove the @p destinationPath if
+ it already exists.
+ */
+- (BOOL)forceMoveItemAtPath:(NSString *)sourcePath
+                     toPath:(NSString *)destinationPath
+                      error:(NSError **)error;
+
 @end