+++ TDTChocolate/FoundationAdditions/NSString+TDTFixedWidthAccessors.h
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+
+@interface NSString (TDTFixedWidthAccessors)
+
+- (int32_t)tdt_int32Value;
+
+- (int64_t)tdt_int64Value;
+
+@end